+++ kosat-web/src/main/kotlin/Client.kt
@@ -1,3 +1,4 @@
+import components.App
 import kotlinx.browser.document
 import react.create
 import react.dom.render
@@ -6,6 +7,6 @@ fun main() {
     val container = document.createElement("div")
     document.body!!.appendChild(container)
 
-    val visualizer = Visualizer.create()
-    render(visualizer, container)
+    val app = App.create()
+    render(app, container)
 }
